requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.0
supabase>=2.0.0
apscheduler>=3.10.0
pyyaml>=6.0
